ci: clippy clean for Rust 1.95 stable
CI's `cargo clippy --workspace --all-targets -- -D warnings` lit up
under the latest stable's lint set (33 errors total across lib +
lib-test + examples). All fixes:

- Unused imports: 7 deletions (build_side_bloom_emitter_exec.rs:
  inline ExecutionPlanProperties re-imports inside test fns that
  the top-level already provided; lane_filter_sum_kernel.rs: Clause).
- Test-only imports relocated: RecordBatch + TryStreamExt moved
  into the `#[cfg(test)] mod tests { use super::*; ... }` block
  since the lib-build doesn't reference them.
- Dead store: `finalized = true;` at the end of the
  BuildSideBloomEmitterExec stream-unfold closure (stream
  terminates immediately after; flag never re-read).
- type_complexity: 4 sites annotated `#[allow]` — the tuple shapes
  carry meaning (build/probe side-emitter state machine, cascading
  primary tracker, RobinHood test-helper return) and aliasing them
  to named types would be churn for no gain.
- too_many_arguments: 3 sites annotated `#[allow]` —
  build_partition_stream_dispatch (Σ.Q.L9 plumbing), push_through_subtree
  (semi-join push), scalar_reference (a bench).
- manual_clamp: 2 sites converted .max(MIN).min(MAX) → .clamp(MIN, MAX).
- Robin-Hood test module: `#[allow(approx_constant, type_complexity)]`
  on the test mod (6 sites of 3.14 in assertions, 3 sites of
  (Vec<(i64, f64, ...)>) return types).
- Useless `&vec![...]` → `&[...]` in cascading-rule test fixture.
- 4 example binaries (sigma_q_l14_q07_sum_check, sigma_q_l15_*,
  profile_query): added missing `require_filtered_build: false`
  field after the L9.SelectiveBuild struct change.
- Dead-fn `stddev` in sigma_q_l13_scan_only_ab.rs: `#[allow(dead_code)]`
  (it's an unused helper retained alongside its mean+sigma siblings).

No semantic changes — every annotation either localises a structural
choice or relocates an import to the only build profile that uses it.

Includes the SF=10 weakness-closure architect doc
(docs/PHASE_SIGMA_T_SF10_WEAKNESS_CLOSURE.md) that landed in parallel.

+++ crates/ematix-flow-core/src/robin_hood_agg.rs
@@ -1326,7 +1326,7 @@ impl RobinHoodSumF64RadixAgg {
         let shift: u32 = if self.radix_bits == 0 {
             0
         } else {
-            (usize::BITS as u32) - self.radix_bits as u32
+            usize::BITS - self.radix_bits as u32
         };
 
         let mut off = 0;
@@ -2381,6 +2381,7 @@ impl TaggedSumF64Agg {
 }
 
 #[cfg(test)]
+#[allow(clippy::type_complexity, clippy::approx_constant)]
 mod tests {
     use super::*;
